pyproject.toml
[tool.poetry]
name = "dockerbot"
version = "0.1.0"
description = "A bot to work with Docker"
authors = ["Lorenzo Carbonell <lorenzo.carbonell.cerezo@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["src"]
defineConstant = { DEBUG = true }

[tool.pytest.ini_options]
log_cli = true
log_level = "DEBUG"
